rename `shared` to `common`
