+++ src/common_serdes/bytes.ts
