+++ src/common_serdes/clazz.ts
