+++ src/common_serdes/size.ts
