+++ src/common_serdes/string.ts
