+++ src/eval_serdes/index.ts
@@ -1,3 +1,3 @@
-export * from "../shared_serdes";
+export * from "../common_serdes";
 export * from "./struct";
 export * from "./number";
