+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from "./shared";
+export * from "./common";
 export * from "./encodings";
 export * from "./serdes";
