+++ src/node_eval.ts
@@ -1,3 +1,3 @@
-export * from "./shared";
+export * from "./common";
 export * from "./node_encodings";
 export * from "./eval_serdes";
